+++ skops/io/tests/test_audit.py
@@ -4,16 +4,13 @@ import re
 from contextlib import suppress
 from zipfile import ZipFile
 
-import numpy as np
 import pytest
 from sklearn.linear_model import LogisticRegression
-from sklearn.pipeline import FeatureUnion, Pipeline
-from sklearn.preprocessing import FunctionTransformer, StandardScaler
 
 from skops.io import dumps, get_untrusted_types
 from skops.io._audit import Node, audit_tree, check_type, get_tree, temp_setattr
-from skops.io._general import DictNode, dict_get_state
-from skops.io._utils import LoadContext, SaveContext, gettype
+from skops.io._general import DictNode, JsonNode, ObjectNode, dict_get_state
+from skops.io._utils import LoadContext, SaveContext, get_state, gettype
 
 
 class CustomType:
@@ -41,31 +38,35 @@ def test_check_type(module_name, type_name, trusted, expected):
 def test_audit_tree_untrusted():
     var = {"a": CustomType(1), 2: CustomType(2)}
     state = dict_get_state(var, SaveContext(None, 0, {}))
-    node = DictNode(state, LoadContext(None), trusted=False)
+    load_context = LoadContext(None, -1)
+
+    node = DictNode(state, load_context, trusted=False)
     with pytest.raises(
         TypeError,
         match=re.escape(
             "Untrusted types found in the file: ['test_audit.CustomType']."
         ),
     ):
-        audit_tree(node, trusted=False)
+        audit_tree(node)
 
     # there shouldn't be an error with trusted=True
-    audit_tree(node, trusted=True)
+    node = DictNode(state, LoadContext(None, -1), trusted=True)
+    audit_tree(node)
 
     untrusted_list = get_untrusted_types(data=dumps(var))
     assert untrusted_list == ["test_audit.CustomType"]
 
     # passing the type would fix it.
-    audit_tree(node, trusted=untrusted_list)
+    node = DictNode(state, LoadContext(None, -1), trusted=untrusted_list)
+    audit_tree(node)
 
 
 def test_audit_tree_defaults():
     # test that the default types are trusted
     var = {"a": 1, 2: "b"}
     state = dict_get_state(var, SaveContext(None, 0, {}))
-    node = DictNode(state, LoadContext(None), trusted=False)
-    audit_tree(node, trusted=[])
+    node = DictNode(state, LoadContext(None, -1), trusted=False)
+    audit_tree(node)
 
 
 @pytest.mark.parametrize(
@@ -97,7 +98,9 @@ def test_list_safety(values, is_safe):
 
     with ZipFile(io.BytesIO(content), "r") as zip_file:
         schema = json.loads(zip_file.read("schema.json"))
-        tree = get_tree(schema, load_context=LoadContext(src=zip_file))
+        tree = get_tree(
+            schema, load_context=LoadContext(src=zip_file, protocol=-1), trusted=False
+        )
         assert tree.is_safe() == is_safe
 
 
@@ -146,20 +149,26 @@ def test_temp_setattr():
     assert not hasattr(temp, "b")
 
 
-def test_complex_pipeline_untrusted_set():
-    # fmt: off
-    clf = Pipeline([
-        ("features", FeatureUnion([
-            ("scaler", StandardScaler()),
-            ("sqrt", FunctionTransformer(
-                    func=np.sqrt,
-                    inverse_func=np.square,
-                )),
-        ])),
-        ("clf", LogisticRegression(random_state=0, solver="liblinear")),
-    ])
-    # fmt: on
+def test_format_object_node():
+    estimator = LogisticRegression(random_state=0, solver="liblinear")
+    state = get_state(estimator, SaveContext(None))
+    node = ObjectNode(state, LoadContext(None, -1))
+    expected = "sklearn.linear_model._logistic.LogisticRegression"
+    assert node.format() == expected
 
-    untrusted = get_untrusted_types(data=dumps(clf))
-    type_names = [x.split(".")[-1] for x in untrusted]
-    assert type_names == ["sqrt", "square"]
+
+@pytest.mark.parametrize(
+    "inp, expected",
+    [
+        ("hello", 'json-type("hello")'),
+        (123, "json-type(123)"),
+        (0.456, "json-type(0.456)"),
+        (True, "json-type(true)"),
+        (False, "json-type(false)"),
+        (None, "json-type(null)"),
+    ],
+)
+def test_format_json_node(inp, expected):
+    state = get_state(inp, SaveContext(None))
+    node = JsonNode(state, LoadContext(None, -1))
+    assert node.format() == expected